backend/requirements.txt
# VISTA Backend Dependencies
# Vision and AI
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
yolox>=0.3.0

# Real-time Communication
python-socketio>=5.8.0
uvicorn>=0.23.0

# Navigation and Maps
requests>=2.31.0

# Speech and Audio
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.11

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
python-dateutil>=2.8.0

# Optional: GPS and Location (uncomment if needed)
# gpsd-py3>=0.3.0
# geopy>=2.3.0

# Optional: Offline Speech Recognition (uncomment if needed)
# pocketsphinx>=5.0.0